+++ blocks/xp/classes/local/leaderboard/leaderboard.php
@@ -15,7 +15,7 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Default shortcodes definition maker.
+ * Leaderboard.
  *
  * @package    block_xp
  * @copyright  2018 Frédéric Massart
@@ -23,44 +23,59 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace block_xp\local\shortcode;
+namespace block_xp\local\leaderboard;
 defined('MOODLE_INTERNAL') || die();
 
+use block_xp\local\sql\limit;
+
 /**
- * Default shortcodes definition maker.
+ * Leaderboard.
  *
  * @package    block_xp
  * @copyright  2018 Frédéric Massart
  * @author     Frédéric Massart <fred@branchup.tech>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class default_shortcodes_definition_maker implements shortcodes_definition_maker {
+interface leaderboard {
 
     /**
-     * Get the shortcodes definition.
+     * Get the leaderboard columns.
      *
-     * @return array
+     * @return array Where keys are column identifiers and values are lang_string objects.
      */
-    public function get_shortcodes_definition() {
-        return [
-            'xpbadge' => [
-                'callback' => 'block_xp\local\shortcode\handler::xpbadge',
-                'description' => 'shortcode:xpbadge'
-            ],
-            'xpiflevel' => [
-                'callback' => 'block_xp\local\shortcode\handler::xpiflevel',
-                'description' => 'shortcode:xpiflevel',
-                'wraps' => true
-            ],
-            'xpladder' => [
-                'callback' => 'block_xp\local\shortcode\handler::xpladder',
-                'description' => 'shortcode:xpladder'
-            ],
-            'xpprogressbar' => [
-                'callback' => 'block_xp\local\shortcode\handler::xpprogressbar',
-                'description' => 'shortcode:xpprogressbar'
-            ],
-        ];
-    }
+    public function get_columns();
+
+    /**
+     * Get the number of entries.
+     *
+     * @return int
+     */
+    public function get_count();
+
+    /**
+     * Return the position of the object.
+     *
+     * The position is used to determine how to paginate the leaderboard.
+     *
+     * @param int $id The object ID.
+     * @return int Indexed from 0, null when not ranked.
+     */
+    public function get_position($id);
+
+    /**
+     * Get the rank of an object.
+     *
+     * @param int $id The object ID.
+     * @return rank|null
+     */
+    public function get_rank($id);
+
+    /**
+     * Get the ranking.
+     *
+     * @param limit $limit The limit.
+     * @return rank[]
+     */
+    public function get_ranking(limit $limit);
 
 }